backend/requirements-dev.txt
# Development tools для статического анализа и форматирования
# Установка: pip install -r requirements-dev.txt

# Code formatting
black==24.10.0
isort==5.13.2

# Linting
ruff==0.8.4  # уже в requirements.txt, но дублируем для ясности

# Type checking
mypy==1.11.2  # уже в requirements.txt
django-stubs==5.0.4
djangorestframework-stubs==3.15.0
types-requests==2.32.0.20240914

# Testing (уже в основном requirements.txt)
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==5.0.0  # добавляем coverage

# Pre-commit hooks (optional)
pre-commit==3.8.0

